Removed dev container. Added android vm install script for linux
