--- React Native/.devcontainer/devcontainer.json
@@ -1,17 +1,0 @@
-{
-    "name": "ReactNative",
-    "dockerComposeFile": "docker-compose-dev.yml",
-    "service": "react-native",
-	"workspaceFolder": "/app",
-	"customizations": {
-		"vscode": {
-			"extensions": [
-				"ms-azuretools.vscode-docker",
-				"ms-vscode-remote.remote-containers",
-				"esbenp.prettier-vscode",
-				"dsznajder.es7-react-js-snippets",
-				"msjsdiag.vscode-react-native"
-			]
-		}
-	}
-}
