+++ React Native/assets/adaptive-icon.png
