+++ React Native/assets/favicon.png
