--- React Native/constants/Layout.ts
@@ -1,12 +1,0 @@
-import { Dimensions } from 'react-native';
-
-const width = Dimensions.get('window').width;
-const height = Dimensions.get('window').height;
-
-export default {
-  window: {
-    width,
-    height,
-  },
-  isSmallDevice: width < 375,
-};
